apps/api/pyproject.toml
[tool.poetry]
name = "blog-api"
version = "0.1.0"
description = "FastAPI backend for tech blog"
authors = ["Junbeom <your-email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
email-validator = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.18"
httpx = "^0.28.1"
authlib = "^1.3.2"
itsdangerous = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
black = "^24.10.0"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
